requirements.txt
# TOON Format Video Demo Requirements
# Install with: pip install -r requirements.txt

# TOON format library (install from GitHub)
git+https://github.com/toon-format/toon-python.git

# Azure OpenAI
openai>=1.0.0

# Token counting
tiktoken>=0.5.0

# Environment variables management
python-dotenv>=1.0.0

# Azure AI Agents (for multi-agent support)
azure-ai-agents>=1.0.0

# Azure Identity (for authentication)
azure-identity>=1.15.0

# Semantic Kernel (for AI orchestration)
semantic-kernel>=1.0.0

# MCP (Model Context Protocol) - required by semantic-kernel's MCP connectors
mcp>=1.22.0

# LangChain (alternative AI orchestration framework)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Agent Framework (for AzureAgentFramework.py)
agent-framework>=0.1.0

# Data validation (for Pydantic models)
pydantic>=2.0.0
